Persistence Context and Work with @Transactional

+++ Hotel1/service-api/src/main/java/com/senlainc/git_courses/java_training/petushok_valiantsin/api/service/IGuestService.java
@@ -9,7 +9,7 @@ public interface IGuestService {
 
     void add(String firstName, String lastName, LocalDate birthday);
 
-    void delete(int index);
+    void delete(long index);
 
     Long num();
